[GOORM-27]-refactor: 패키지명 변경
