+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'someone'
+rootProject.name = 'familynote'
