+++ src/main/java/goorm/kgu/familynote/FamilyNoteApplication.java
@@ -1,0 +1,13 @@
+package goorm.kgu.familynote;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FamilyNoteApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FamilyNoteApplication.class, args);
+	}
+
+}
