+++ src/main/java/goorm/kgu/familynote/common/exception/CustomException.java
@@ -1,0 +1,19 @@
+package goorm.kgu.familynote.common.exception;
+
+import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+	private final ExceptionCode code;
+
+	public CustomException(ExceptionCode code) {
+		super(code.getMessage());
+		this.code = code;
+	}
+
+	public boolean isServerError() {
+		return code.getStatus().equals(INTERNAL_SERVER_ERROR);
+	}
+}
