+++ src/main/java/goorm/kgu/familynote/common/exception/ExceptionCode.java
@@ -1,0 +1,11 @@
+package goorm.kgu.familynote.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionCode {
+	HttpStatus getStatus();
+
+	String getCode();
+
+	String getMessage();
+}
