+++ src/main/java/goorm/kgu/familynote/domain/user/domain/User.java
@@ -1,0 +1,30 @@
+package goorm.kgu.familynote.domain.user.domain;
+
+import goorm.kgu.familynote.common.domain.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Builder
+@Table(name = "\"user\"")
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User extends BaseTimeEntity {
+
+	@Id
+	@Column(nullable = false)
+	private String id;
+
+	@Column(nullable = false)
+	private String name;
+
+	private String email;
+}
