+++ src/main/java/goorm/kgu/familynote/domain/user/domain/UserRepository.java
@@ -1,0 +1,4 @@
+package goorm.kgu.familynote.domain.user.domain;
+
+public interface UserRepository {
+}
