+++ src/main/java/goorm/kgu/familynote/domain/user/presentation/request/UserRequest.java
@@ -1,0 +1,8 @@
+package goorm.kgu.familynote.domain.user.presentation.request;
+
+public record UserRequest(
+	String name,
+	String email,
+	String password
+) {
+}
