+++ src/main/java/goorm/kgu/familynote/domain/user/presentation/response/UserResponse.java
@@ -1,0 +1,8 @@
+package goorm.kgu.familynote.domain.user.presentation.response;
+
+public record UserResponse(
+	Long id,
+	String name,
+	String email
+) {
+}
