--- src/main/java/goorm/kgu/someone/SomeoneApplication.java
@@ -1,13 +1,0 @@
-package goorm.kgu.someone;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SomeoneApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SomeoneApplication.class, args);
-	}
-
-}
