--- src/main/java/goorm/kgu/someone/common/exception/ExceptionCode.java
@@ -1,11 +1,0 @@
-package goorm.kgu.someone.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-	HttpStatus getStatus();
-
-	String getCode();
-
-	String getMessage();
-}
