--- src/main/java/goorm/kgu/someone/common/logging/LoggingAspect.java
@@ -1,58 +1,0 @@
-package goorm.kgu.someone.common.logging;
-
-import java.lang.reflect.Method;
-import java.util.Arrays;
-
-import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.annotation.AfterThrowing;
-import org.aspectj.lang.annotation.Around;
-import org.aspectj.lang.annotation.Aspect;
-import org.aspectj.lang.annotation.Pointcut;
-import org.aspectj.lang.reflect.MethodSignature;
-import org.springframework.stereotype.Component;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Aspect
-@Component
-public class LoggingAspect {
-
-	@Pointcut(
-		"execution(* goorm.kgu.someone.domain..api.*Controller.*(..)) || " +
-			"execution(* goorm.kgu.someone.domain..api.*.*Controller.*(..))"
-	)
-	private void cut() {
-	}
-
-	@Around("cut()")
-	public Object around(ProceedingJoinPoint joinPoint) throws Throwable {
-		long startTime = System.currentTimeMillis();
-
-		Method method = getMethod(joinPoint);
-		log.info("Method Log: {} || Args: {}", method.getName(), Arrays.toString(joinPoint.getArgs()));
-
-		Object result = joinPoint.proceed();
-
-		long endTime = System.currentTimeMillis();
-		long executionTime = endTime - startTime;
-
-		log.info("Method {} is finished || Execution Time: {} ms", method.getName(), executionTime);
-		return result;
-	}
-
-	@AfterThrowing(pointcut = "cut()", throwing = "exception")
-	public void afterThrowing(JoinPoint joinPoint, Throwable exception) {
-		Method method = getMethod(joinPoint);
-		log.error("AfterThrowing Method: {} || Exception: {}", method.getName(), exception.getMessage());
-		log.error("Exception type: {}", exception.getClass().toGenericString());
-		log.error("Exception final point: {}", exception.getStackTrace()[0]);
-		log.error("Exception point: {}", exception.getStackTrace()[1]);
-	}
-
-	private Method getMethod(JoinPoint joinPoint) {
-		MethodSignature signature = (MethodSignature)joinPoint.getSignature();
-		return signature.getMethod();
-	}
-}
