--- src/main/java/goorm/kgu/someone/domain/user/domain/UserRepository.java
@@ -1,4 +1,0 @@
-package goorm.kgu.someone.domain.user.domain;
-
-public interface UserRepository {
-}
