--- src/main/java/goorm/kgu/someone/domain/user/infrastructure/JpaUserRepository.java
@@ -1,8 +1,0 @@
-package goorm.kgu.someone.domain.user.infrastructure;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import goorm.kgu.someone.domain.user.domain.User;
-
-public interface JpaUserRepository extends JpaRepository<User, Long>{
-}
