--- src/main/java/goorm/kgu/someone/domain/user/presentation/UserController.java
@@ -1,16 +1,0 @@
-package goorm.kgu.someone.domain.user.presentation;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import goorm.kgu.someone.domain.user.application.UserService;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/users")
-@Tag(name = "User", description = "유저 관련 api / 담당자 : 이한음")
-public class UserController {
-	private final UserService userService;
-}
