--- src/main/java/goorm/kgu/someone/domain/user/presentation/request/UserRequest.java
@@ -1,8 +1,0 @@
-package goorm.kgu.someone.domain.user.presentation.request;
-
-public record UserRequest(
-	String name,
-	String email,
-	String password
-) {
-}
