--- src/main/java/goorm/kgu/someone/domain/user/presentation/response/UserResponse.java
@@ -1,8 +1,0 @@
-package goorm.kgu.someone.domain.user.presentation.response;
-
-public record UserResponse(
-	Long id,
-	String name,
-	String email
-) {
-}
