+++ src/test/java/goorm/kgu/familynote/FamilyNoteApplicationTests.java
@@ -1,0 +1,13 @@
+package goorm.kgu.familynote;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FamilyNoteApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
