--- src/test/java/goorm/kgu/someone/SomeoneApplicationTests.java
@@ -1,13 +1,0 @@
-package goorm.kgu.someone;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SomeoneApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
